Conditionalize overwriting of config/database.yml in empirical:setup
Keep customized local config/database.yml if it already exists when running
rake empirical:setup. i.e. do not overwrite it with the example database.yml.

Without this, the database credentials would never be present when required for
db:create, when subsequently called in empirical:setup.

Consider two scenarios:
1. Clean install.
2. Existing install, when empirical:setup called to refresh database.

Scenario 1:
- config/database.yml.example needs to be customized with database credentials.
- Accordingly, between copying it to config/database.yml and db:create, the local
  developer needs to modify it to add credentials.
- Thus, current approach of always overwriting config/database.yml will never work.
- All documentation states that prior to running rake empirical:setup, the example
  config/database.yml.example should be copid to config/database.yml and modified.
- Thus, (a) the developer has already done so, in which case it shouldn't be
  overwritten with the example database.yml; or (b) it hasn't been done so, and then
  copying the example database.yml is the best we can do (will still fail at db:create
  as no database credentials provided, but at least it means the example database.yml
  is in the right location for the developer to customize and then rerun
  rake empirical:setup).

Scenario 2:
- config/database.yml has already been customized with database credentials.
- No reason to overwrite it with the example database.yml

Partially addresses: https://github.com/empirical-org/Empirical-Core/issues/1811

+++ lib/tasks/setup.rake
@@ -5,8 +5,10 @@ namespace :empirical do
     puts "** Creating tmp directories..."
     Rake::Task["tmp:create"].invoke
 
-    puts "** Copying DB Credentials..."
-    `cp config/database.yml.example config/database.yml`
+    unless File.exist?("config/database.yml")
+      puts "** Copying DB Credentials..."
+      `cp config/database.yml.example config/database.yml`
+    end
 
     puts '** Copying env variables...'
     `cp .env-sample .env`
